PR-SUG: use message extension instead to retrieve correlation info
