+++ src/Arcus.Messaging.ServiceBus.Core/Extensions/MessageExtensions.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using Arcus.Messaging.Abstractions;
-using GuardNet;
 
 // ReSharper disable once CheckNamespace
 namespace Microsoft.Azure.ServiceBus
@@ -30,20 +28,6 @@ namespace Microsoft.Azure.ServiceBus
 
             throw new KeyNotFoundException(
                 $"No user property with the key: '{key}' was found in the Service Bus message");
-        }
-
-        /// <summary>
-        ///     Gets the transaction id that is linked to this message
-        /// </summary>
-        /// <param name="message">Message to process</param>
-        /// <returns>Transaction id for message</returns>
-        public static string GetTransactionId(this Message message)
-        {
-            Guard.NotNull(message, nameof(message));
-
-            return message.UserProperties.TryGetValue(PropertyNames.TransactionId, out object transactionId)
-                ? transactionId.ToString()
-                : string.Empty;
         }
     }
 }
